pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zabbix-mcp"
version = "0.1.0"
description = "Zabbix MCP read-only query service"
readme = ""
requires-python = ">=3.8"
license = { text = "Apache-2.0" }
authors = [{ name = "Zabbix-MCP Authors" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.31.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "typer>=0.12.5",
    "rich>=13.9.0",
    "prometheus-client>=0.20.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
zabbix-mcp = "zabbix_mcp.cli:app"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"

